moved everything to src folder
