+++ src/static/images/.gitignore
@@ -1,4 +1,4 @@
 # Ignore everything in this directory
-*
+.gitignore
 # Except this file
 !.gitignore
