+++ src/static/layouts/.gitignore
@@ -1,4 +1,4 @@
 # Ignore everything in this directory
-*
+besu2.json
 # Except this file
 !.gitignore
